pyproject.toml
[project]
name = "chatbots-convo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "MuhammadRaffey", email = "muhammadraffey26@gmail.com" }]
requires-python = ">=3.11"
dependencies = ["openai>=1.63.0", "python-dotenv>=1.0.1"]

[project.scripts]
demo = "chatbots_convo.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
